Client sends a distance and save it in DB

+++ server/src/main/java/com/autonomoussystemserver/server/controller/model/DistanceDto.java
@@ -3,9 +3,9 @@ package com.autonomoussystemserver.server.controller.model;
 public class DistanceDto {
     private Integer fromDevice;
     private Integer toDevice;
-    private Integer distance;
+    private Long distance;
 
-    public DistanceDto(Integer fromDevice, Integer toDevice, Integer distance) {
+    public DistanceDto(Integer fromDevice, Integer toDevice, Long distance) {
         this.fromDevice = fromDevice;
         this.toDevice = toDevice;
         this.distance = distance;
@@ -19,7 +19,7 @@ public class DistanceDto {
         return toDevice;
     }
 
-    public Integer getDistance() {
+    public Long getDistance() {
         return distance;
     }
 }
